+++ CHANGELOG.md
@@ -12,6 +12,7 @@ github compare-link with the previous one.
 ## [Unreleased](https://github.com/asdf-community/asdf-direnv/compare/v0.3.0..master)
 
 - Add support for resolving 'latest:X' in .tool-versions #136
+- Add support for resolving 'latest' in .tool-versions #168
 
 ## [0.3.0](https://github.com/asdf-community/asdf-direnv/compare/v0.3.0) (2022-04-03)
 
@@ -27,7 +28,7 @@ github compare-link with the previous one.
 
 - Speed up direnv stdlib by trying not to use asdf exec. #120
 
-- Do not add plugin paths for "system" versions. #116 
+- Do not add plugin paths for "system" versions. #116
 
 
 ## [0.2.0](https://github.com/asdf-community/asdf-direnv/releases/v0.2.0) (2022-03-16)